изменено перенаправление регистрация - логин, логин - мои задачи изменено расположение и видимость кнопок, если пользователь авторизован - убирается главная, о нас, гайд по заданиям добавлена валидация даты добавлен декоратор @login_required к форуму и созданию задач - чтобы что-либо писать нужна авторизация

+++ todolist/tasks/views.py
@@ -3,6 +3,7 @@ from django.shortcuts import render, redirect, get_object_or_404
 from django.contrib.auth.decorators import login_required
 from .forms import TaskForm
 from .models import Task
+from django.contrib import messages
 
 def index(request):
     context = {
@@ -24,17 +25,19 @@ def guide(request):
     }
     return render(request, 'tasks/guide.html', context)
 
-@login_required
+
+@login_required()
 def create_task(request):
     if request.method == 'POST':
         form = TaskForm(request.POST)
-        task = form.save(commit=False)
         if form.is_valid():
+            task = form.save(commit=False)
             task.user = request.user
-            form.save()
+            task.save()
             return redirect('tasks:mytasks')  # Перенаправление на страницу со списком задач
     else:
         form = TaskForm()
+
     context = {
         'form': form,
     }
@@ -43,20 +46,31 @@ def create_task(request):
 @login_required
 def edit_task(request, task_id):
     task = get_object_or_404(Task, id=task_id)
+
     if task.user != request.user:
         return redirect('tasks:mytasks')  # Перенаправляем, если пользователь не владелец
 
     if request.method == 'POST':
+        if 'undone' in request.POST:
+            task.status = False
+            task.save()
+            form = TaskForm(request.POST, instance=task) # Перезагружаем форму, чтобы данные не сбрасывались
+
+            if form.is_valid():
+                form.save()  # Сохраняем данные формы
+            return redirect('tasks:mytasks')
+
         if 'delete' in request.POST:
             task.delete()
-            return redirect('tasks:mytasks')  # Перенаправление после удаления
-        else:
-            form = TaskForm(request.POST, instance=task)
-            if form.is_valid():
-                form.save()
-                return redirect('tasks:mytasks')  # Перенаправление после редактирования
+            return redirect('tasks:mytasks')
+
+        form = TaskForm(request.POST, instance=task)
+        if form.is_valid():
+            form.save()
+            return redirect('tasks:mytasks')
     else:
         form = TaskForm(instance=task)
+
     context = {
         'form': form,
         'task': task,
@@ -71,4 +85,12 @@ def mytasks(request):
         'tasks': tasks
                }
     return render(request, 'tasks/mytasks.html', context)
+
+
+def completed_task(request, task_id):
+    task = get_object_or_404(Task, id=task_id, user=request.user)
+    if not task.status:  # Если задача еще не выполнена
+        task.status = True
+        task.save()
+    return redirect('tasks:mytasks')
